vectorizer-worker/pyproject.toml
[project]
name = "vectorizer-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.14",
    "aiokafka>=0.12.0",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "confluent-kafka==2.11.0",
    "cryptography==45.0.5",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "faiss-cpu==1.11.0.post1",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "greenlet==3.2.3",
    "groq==0.30.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.33.4",
    "idna==3.10",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kafka-python==2.2.15",
    "langchain==0.3.26",
    "langchain-community==0.3.27",
    "langchain-core==0.3.71",
    "langchain-groq==0.3.6",
    "langchain-huggingface==0.3.1",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.4.8",
    "lxml==6.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mpmath==1.3.0",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "networkx==3.5",
    "numpy==2.3.1",
    "orjson==3.11.0",
    "packaging==25.0",
    "pdfminer-six==20250506",
    "pdfplumber==0.11.7",
    "pillow==11.3.0",
    "propcache==0.3.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pypdf==5.8.0",
    "pypdfium2==4.30.0",
    "python-docx==1.2.0",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "safetensors==0.5.3",
    "scikit-learn==1.7.1",
    "scipy==1.16.0",
    "sentence-transformers==5.0.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "starlette==0.47.2",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.2",
    "torch==2.7.1",
    "tqdm==4.67.1",
    "transformers==4.53.3",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
